+++ scripts/clean_tyndp_benchmark.py
@@ -246,9 +246,35 @@ def load_benchmark(
             .groupby(["scenario", "year", "unit", "table"])
             .sum()
             .reset_index()
-            .assign(carrier="total (excl. SMR and power generation)")
+            .assign(carrier="exogenous demand")
         )
         df_group = df_converted[df_converted.carrier.isin(group)]
+        df_converted = pd.concat([df_group, df_other], ignore_index=True)
+    # Aggregate hydrogen demand to match input data resolution
+    elif table == "hydrogen_demand":
+        group = ["power generation", "e-fuels"]
+        df_other = (
+            df_converted[
+                (~df_converted.carrier.isin(group)) & (df_converted.carrier != "total")
+            ]
+            .groupby(["scenario", "year", "unit", "table"])
+            .sum()
+            .reset_index()
+            .assign(carrier="exogenous demand")
+        )
+        df_group = df_converted[df_converted.carrier.isin(group)]
+        df_converted = pd.concat([df_group, df_other], ignore_index=True)
+    # Aggregate hydrogen import sources together to match input data resolution
+    elif table == "hydrogen_supply":
+        to_group = ["low carbon imports", "renewable imports"]
+        df_other = (
+            df_converted[df_converted.carrier.isin(to_group)]
+            .groupby(["scenario", "year", "unit", "table"])
+            .sum()
+            .reset_index()
+            .assign(carrier="imports (renewable & low carbon)")
+        )
+        df_group = df_converted[~df_converted.carrier.isin(to_group)]
         df_converted = pd.concat([df_group, df_other], ignore_index=True)
     # Remove carriers not explicitly modeled
     elif table == "final_energy_demand":